+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/primitive/repair/ConstraintGroupBasedRepair.java
@@ -154,7 +154,7 @@ public class ConstraintGroupBasedRepair implements OnlineOptimization {
                     .map(group -> constraint
                             .lineProcessing()
                             .columnView(INCOMING_CONSTRAINT_GROUP)
-                            .lookup(group)
+                            .persistedLookup(group)
                             .columnView(LINE)
                             .values())
                     .flatMap(streamOfLineList -> streamOfLineList.stream())
@@ -164,7 +164,7 @@ public class ConstraintGroupBasedRepair implements OnlineOptimization {
                             return !constraint
                                     .lineProcessing()
                                     .columnView(LINE)
-                                    .lookup(allocation)
+                                    .persistedLookup(allocation)
                                     .unorderedLines()
                                     .get(0)
                                     .value(RATING)