+++ projects/net.splitcells.gel.ext/src/main/java/net/splitcells/gel/data/lookup/LookupTest.java
@@ -45,37 +45,37 @@ public class LookupTest extends TestSuiteI {
         final var b = attribute(Integer.class, "b");
         final var testSubject = tableI(a, b);
         testSubject.addTranslated(list(1, 2));
-        testSubject.columnView(a).lookup(1).unorderedLines().requireSizeOf(1);
-        testSubject.columnView(a).lookup(3).unorderedLines().requireEmpty();
-        testSubject.columnView(a).lookup(1).columnView(b).lookup(2)
+        testSubject.columnView(a).persistedLookup(1).unorderedLines().requireSizeOf(1);
+        testSubject.columnView(a).persistedLookup(3).unorderedLines().requireEmpty();
+        testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(2)
                 .unorderedLines()
                 .requireSizeOf(1);
         requireNotNull(testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .rawLinesView()
                 .get(0));
         testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(3)
+                .persistedLookup(3)
                 .unorderedLines()
                 .requireEmpty();
         testSubject
                 .columnView(a)
-                .lookup(3)
+                .persistedLookup(3)
                 .columnView(b)
-                .lookup(3)
+                .persistedLookup(3)
                 .unorderedLines()
                 .requireEmpty();
         testSubject
                 .columnView(a)
-                .lookup(3)
+                .persistedLookup(3)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .unorderedLines()
                 .requireEmpty();
     }
@@ -93,54 +93,54 @@ public class LookupTest extends TestSuiteI {
         testSubject.addTranslated(list(3, 3));
         testSubject.addTranslated(list(1, 3));
         testSubject.addTranslated(list(3, 2));
-        testSubject.columnView(a).lookup(1).unorderedLines().requireSizeOf(3);
+        testSubject.columnView(a).persistedLookup(1).unorderedLines().requireSizeOf(3);
         testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .unorderedLines()
                 .requireSizeOf(2);
         requireNotNull(testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .rawLinesView()
                 .get(0));
         requireNotNull(testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .rawLinesView()
                 .get(1));
         testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(3)
+                .persistedLookup(3)
                 .unorderedLines()
                 .requireSizeOf(1);
         testSubject
                 .columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(4)
+                .persistedLookup(4)
                 .unorderedLines()
                 .requireEmpty();
         testSubject
                 .columnView(a)
-                .lookup(3)
+                .persistedLookup(3)
                 .columnView(b)
-                .lookup(2)
+                .persistedLookup(2)
                 .unorderedLines()
                 .requireSizeOf(1);
         testSubject
                 .columnView(a)
-                .lookup(3)
+                .persistedLookup(3)
                 .columnView(b)
-                .lookup(1)
+                .persistedLookup(1)
                 .unorderedLines()
                 .requireEmpty();
     }
@@ -154,22 +154,22 @@ public class LookupTest extends TestSuiteI {
         {
             testSubject
                     .columnView(a)
-                    .lookup(1)
+                    .persistedLookup(1)
                     .columnView(b)
-                    .lookup(1)
+                    .persistedLookup(1)
                     .unorderedLines()
                     .requireEmpty();
             testSubject
                     .columnView(a)
-                    .lookup(1)
+                    .persistedLookup(1)
                     .columnView(b)
-                    .lookup(2)
+                    .persistedLookup(2)
                     .unorderedLines()
                     .requireEmpty();
-            testSubject.columnView(a).lookup(1).columnView(b).lookup(3).unorderedLines().requireEmpty();
-            testSubject.columnView(a).lookup(1).columnView(b).lookup(4).unorderedLines().requireEmpty();
-            testSubject.columnView(a).lookup(3).columnView(b).lookup(2).unorderedLines().requireEmpty();
-            testSubject.columnView(a).lookup(3).columnView(b).lookup(1).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(3).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(4).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(3).columnView(b).persistedLookup(2).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(3).columnView(b).persistedLookup(1).unorderedLines().requireEmpty();
         }
         {
             testSubject.addTranslated(list(1, 2));
@@ -181,14 +181,14 @@ public class LookupTest extends TestSuiteI {
             testSubject.addTranslated(list(3, 2));
         }
         {
-            testSubject.columnView(a).lookup(1).unorderedLines().requireSizeOf(3);
-            testSubject.columnView(a).lookup(1).columnView(b).lookup(2).unorderedLines().requireSizeOf(2);
-            requireNotNull(testSubject.columnView(a).lookup(1).columnView(b).lookup(2).rawLinesView().get(0));
-            requireNotNull(testSubject.columnView(a).lookup(1).columnView(b).lookup(2).rawLinesView().get(1));
-            testSubject.columnView(a).lookup(1).columnView(b).lookup(3).unorderedLines().requireSizeOf(1);
-            testSubject.columnView(a).lookup(1).columnView(b).lookup(4).unorderedLines().requireEmpty();
-            testSubject.columnView(a).lookup(3).columnView(b).lookup(2).unorderedLines().requireSizeOf(1);
-            testSubject.columnView(a).lookup(3).columnView(b).lookup(1).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(1).unorderedLines().requireSizeOf(3);
+            testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(2).unorderedLines().requireSizeOf(2);
+            requireNotNull(testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(2).rawLinesView().get(0));
+            requireNotNull(testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(2).rawLinesView().get(1));
+            testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(3).unorderedLines().requireSizeOf(1);
+            testSubject.columnView(a).persistedLookup(1).columnView(b).persistedLookup(4).unorderedLines().requireEmpty();
+            testSubject.columnView(a).persistedLookup(3).columnView(b).persistedLookup(2).unorderedLines().requireSizeOf(1);
+            testSubject.columnView(a).persistedLookup(3).columnView(b).persistedLookup(1).unorderedLines().requireEmpty();
         }
     }
 
@@ -200,9 +200,9 @@ public class LookupTest extends TestSuiteI {
         final var testSubject = tableI(a, b);
         testSubject.addTranslated(list(1, 2));
         final Predicate<Integer> predicate = aa -> aa == 1;
-        testSubject.columnView(a).lookup(1).unorderedLines().requireSizeOf(1);
-        testSubject.columnView(a).lookup(3).unorderedLines().requireEmpty();
-        testSubject.columnView(a).lookup(predicate).unorderedLines().requireSizeOf(1);
+        testSubject.columnView(a).persistedLookup(1).unorderedLines().requireSizeOf(1);
+        testSubject.columnView(a).persistedLookup(3).unorderedLines().requireEmpty();
+        testSubject.columnView(a).persistedLookup(predicate).unorderedLines().requireSizeOf(1);
     }
 
     @Test
@@ -211,9 +211,9 @@ public class LookupTest extends TestSuiteI {
         final var b = attribute(Integer.class, "b");
         final var testSubject = Tables.table(list(a, b), list());
         final Supplier<View> testValues = () -> testSubject.columnView(a)
-                .lookup(1)
+                .persistedLookup(1)
                 .columnView(b)
-                .lookup(2);
+                .persistedLookup(2);
         testValues.get().unorderedLines().requireEmpty();
         final var firstLine = testSubject.addTranslated(list(1, 1));
         testValues.get().unorderedLines().requireEmpty();